requirements.txt
requests
praw
pandas
numpy
beautifulsoup4
python-dotenv
jupyter
matplotlib
seaborn